revert: move logging + observability back to internal/
PR #7 lifted internal/logging and internal/observability to public
packages so cyoda-go-cassandra's full binary could use the same init
helpers. That was an unplanned scope addition on a migration that is
supposed to preserve cyoda-light-go's public/internal split. The
duplication cost of inlining the ~40 lines of init logic in each
downstream binary is trivial; keeping the public API surface narrow
matters more.

Downstream binaries (cyoda-go-cassandra) will inline their own slog
setup + observability.Init equivalent. The inline code is small,
stable, and doesn't create a public-API contract that needs
maintenance.
